moving backend related stuff to ./backend folder
